requirements.txt
pygraphviz
slack_sdk
openai
langchain
